[keymgr] Bump KDFMaxWidth to allow another rom_ctrl connection
We've got an assertion (AdvDataWidth_A) that checks AdvDataWidth isn't
bigger than the MAC operation that we're happy to do. With a single
KDFMaxWidth=1600 was a sensible upper bound.

This isn't enough with an extra rom_ctrl port! (We end up with
AdvDataWidth=1920)

Bump KDFMaxWidth by a small amount to make everything fit again.

Change-Id: Ia9cb1b189574fdf61565451f91a346410d772c93
Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>