[reggen] Pass idx0/count instead of min/max idx to Field.make_multi
No functional change, but this finishes the job of removing
information about "registers" from Field.make_multi (and it makes the
code very slightly simpler).

Since I was touching the lines anyway, I've also removed a
special-case where we didn't copy self.bits if the offset was zero. I
think this was probably a misguided "optimisation" from me in 2021!

I've also reflowed a long line where we construct the Field objects
for the returned list.

Change-Id: Ia91c05dadcf76449705ea5c0987739d052a12b36
Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>