[opentitanlib] Handle missing ownership seal in json
The ownership seal is a value applied to the ownership config by the
chip. It is meaningless to construct an owner config with a seal an
therefore, it should be permitted to omit it from the JSON
representation.

When decoding and printing an ownership config from a chip, however, it
is useful to print the value of the seal.

Change-Id: Ibfc941df60268ff8c6de0b4f9eb2a380567f29e0
Signed-off-by: Chris Frantz <cfrantz@google.com>
(cherry picked from commit e61786d9d1fe27d82f8f089a154768b2a5a2bcfa)
(cherry picked from commit 95b1ec85a9e049b7df4be67f5d2e06aecbc048e0)