[bazel] Rework HW files targets
This commit switches the HW targets (verilator, bitstream) to use
the target create by `opentitan_top`. This allows to remove many of
the hacky `all_files` targets used everywhere. This commit does keep
a few and even creates some, but tries to adhere to the following
convention: only use `glob()` or `all_files` for sub-directories
that contain no bazel targets and should be passed "as-is" to
fusesoc.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>