[otp] Remove SiVal manuf identifier from emulation images.
The use of `CREATOR_SW_CFG_MANUF_STATE` as the SKU identifier is getting
deprecated in favor of using the DeviceID. The DeviceID definition
contains enough information to constraint code signatures to a
particular SKU.

The `CREATOR_SW_CFG_MANUF_STATE` will be used to capture the
manufacturing state of the device. At the end of manufacturing, there
will be a value configured to flag end of manufacturing.

Change-Id: Ic08bd7d9b229cadd2989cf9aef02a636938854ad
Signed-off-by: Miguel Osorio <miguelosorio@google.com>