[sival, keymgr] Fix assertion failure in keymgr key derivation vseq
`chip_sw_keymgr_key_derivation_vseq` provides backdoor support
required in the `chip_sw_keymgr_key_derivation` software test.
However, the test would trigger an assertion failure when run in DV.

The cause of this is a missing space character in the log message that
is used as a checkpoint in the sequence.

Signed-off-by: Andrea Caforio <andrea.caforio@lowrisc.org>