[ci] build and cache bitstream pre-merge with merge queue
With this change and merge queue enabled, PRs will only be merged to
target branch once the bitstream is built and cached. This prevent RTL
changes from causing all subsequent CI jobs to rebuild bitstream after
merging and before bitstreams are built and cached.

This also has the advantage of reducing likelihood of merge skews, when
we move jobs from push to merge_group event. These are future
possibilities, for now only the merge queue is only gated on bitstream
being built.

Change-Id: If1b8d100fa49ed6a1e49d6d3b2d725aa999a0b1a
Signed-off-by: Gary Guo <gary.guo@lowrisc.org>