[crypto] Consolidate P-384 OTBN binaries.
Analogous to the P-256 refactor but more dramatic, because P-384 was split into
more pieces previously. With larger IMEM, we can use only one binary for P-384
operations, instead of the previous four (keygen/sign/verify/ecdh).

There's again a lot of code churn here, but the vast majority of it is only
moved, not changed. Like with P-256, some names had to be changed to fit the
naming convention of definitions from the same C header starting with the same
prefix (so e.g. "ecdsa_p384" became "p384_ecdsa").

This saves about 16.4 kB (!) of code size.

Signed-off-by: Jade Philipoom <jadep@zerorisc.com>