[rules] Add support for preprocessing linker script files
With this commit, linker script files are now preprocessed.
The is handled similarly to `cc_library`: if `ld_library` is used
with only files in `includes`, then it simply adds them to
the compilation context. If `ld_library` is used with `script` set,
then the file with be preprocessed with the compilation context
constructed from its dependencies.

This commit also adds a `defines` attribute to the `ld_library` so
that defines can be set in bazel BUILD files.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>