[silicon_creator] Make the `nonce.h` header available separately
The `nonce.c` file has some requirements that pull a lot of code.
On the other hand, some headers only require the definition of
a nonce and not the implementation that goes with it. This commit
refactors the dependencies so that it is possible to depend on
the nonce and boot data definition without depending on the
flash_ctrl driver.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>