[util,tlgen] Make --outdir a required argument and exit if empty
Before this change, you get a confusing error message if you don't
supply it. The original author also obviously imagined that
log.error() would end the program, but it doesn't! Bail properly if
needed.

Finally, tighten up the test with is_dir() because Path('') gives the
current directory (which is a directory...)

Change-Id: Iaf124af14a3bf0427887cda23560d67f65d9d22c
Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>