[manuf] Make orchestrator resolve paths using runfiles library
This simplifies use of the packaged orchestrator script. Instead of
having to extract the packaged runfiles manually, the rules_python
machinery extracts it into a temp dir and the library handles path
resolution.

Some hackery was required to preserve the scheme used previously for
referring to the main repo vs external repos.

Change-Id: I4072130d6fcaf645f6cefff2df092c40551f7afe
Signed-off-by: Noah Moroze <noah@opentitan.org>