[pwm] Logic simplifications
Reduce the expense of the pos|neg_htbt and pattern_repeat determination;
this is not a functional change, it just requires fewer gates.
Rename the 'htbt_direction' signal to be more meaningful.
Simplify the way the mode-dependent duty cycle ('duty_cycle_actual') is
selected.
Introduce an omitted reset ('dc_wrap_q') in the heartbeat logic when the
channel parameters are changed.

Signed-off-by: Adrian Lees <a.lees@lowrisc.org>