[bazel] resolve conflicting actions in `update_manifest`
The `update_manifest` function was previously using the same output file
name, which is the name of the updated manifest, for different ELF files,
resulting in conflicting actions when multiple environments are set in
`exec_env`.

This modifies the output file name to include the name of the execution
environment, ensuring unique output file names for each ELF file and
resolving the conflict.

Signed-off-by: Anthony Chen <antchen@google.com>