[dv,pwm] Tidy monitor and scoreboard prediction
Duty cycle prediction was very overcomplicated, confusing the task
of forming the current prediction with that of then updating the
state for the subsequent prediction.

Tidy and clarify the detection of active and inactive output states
within the monitor.

The currently-required 'settle time' at start up and transition
between blinking phases may be achieved using the same logic.

Signed-off-by: Adrian Lees <a.lees@lowrisc.org>