[hmac,dv] Skip msg_length check when reset
- when a reset a is happening, the HMAC DV env is not really robust.
Some tasks are still ongoing and would require to be killed. A temporary
fix has been proposed to skip the msg_length check when a Save and
Restore task is ongoing when a reset is triggered. A better approach
will be put in place later when a proper reset agent will be used.

Signed-off-by: Martin Velay <mvelay@lowrisc.org>