[util,otp] Make OTP Image MEM lines deterministic
Sorts the set of word annotations for a MEM line in the OTP image, such
that the MEM line being read by Vivado is the same, but the word
annotations always appear in the same order, improving the determinism
of the builds.

Previously, usage of Python's sets meant that for MEM lines with multiple
word attributes, they could appear in different orders between builds,
causing slight differences in the output images (and thus potentially
requiring later actions to be re-run instead of cache hits due to
differences in the generated OTP image).

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>