[ci] Remove default `local = True` flag from OT tests
As per https://bazel.build/reference/be/general#genrule.local, Bazel
normally runs tests with a default of `local = False`. Setting `local =
True` as we do for FPGA and Verilator tests by default forces genrule
to use the "local" spawn strategy, which disallows remote execution,
sandboxing and remote caching.

We do not use remote execution, but there is no reason for our FPGA /
Verilator tests to be run without sandboxing or without using remotely
cached test results. Bazel will still respect the `cache_test_results`
option, and so use of this cache can be overridden. Any tests that need
to opt out of sandboxing features as the tests originally did can either
add the `no-sandbox` tag to remove sandboxing, or the `local` tag to
replicate the original functionality, though this will meant that these
tests cannot be remotely cached and thus must be re-run every time if
there is no result in a local cache.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>