[rom_ctrl,dv] Properly wait for outputs to avoid race
This avoids a rather confusing race in rom_ctrl_scoreboard where we
were waiting for a change in pwrmgr_data or keymgr_data and then doing
something with the result.

Unfortunately, the structure that we waited on got updated by multiple
processes and they happened to be scheduled with the scoreboard in the
middle.

Use a clocking block to sort things out.

Change-Id: I19500aedefab95b8e2d12c626dc9027cf6374b52
Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>