[hw,sram_ctrl,rtl] Add flop stage on RAM output
This PR adds support for an optional flop stage on the RAM
macro output. The support for this stage is already present
but this PR adds the necesary parameters to sram_ctrl to be
able to configure it. For Darjeeling, this flop stage is
enabled on all sram_ctrl isntances. For the other tops it
remains disabled.

Signed-off-by: Robert Schilling <rschilling@rivosinc.com>