[tools,topgen] Rename clock attribute of reset nodes
This makes the complete top config hjson have the same names
for attributes. Previously the original name for this was "clk"
but the complete config renames it to "clock" because of the
Clocks class field names. The Clocks class is used quite a bit
more in topgen, so this rename creates less canges, and it is
arguably a better name.

Part of #25920

Signed-off-by: Guillermo Maturana <maturana@opentitan.org>