sphincsplus: accept PUB key as PEM encoded ASN.1 objects
Some HSMs distribute SPX pub keys as PEM encoded ANS.1 objects. The
approach so far has been to manually extract the 32 byte key material
from the object. This patch adds parsing code for processing the ANS.1
key files.

The supported algorithm is slh-dsa-sha2-128s-with-sha256, its OID is
defined in
https://csrc.nist.gov/projects/computer-security-objects-register/algorithm-registration#heading1725030685275_13

Tested on an opentitatn owner's firmware by passing to --spx-key first
the raw SPX key and then the HSM generated PEM file, and observing the
unchanged contents of the resulting image.

Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>