[dttool] Move most of the logic out of the templates and into helpers
The templates for the various files were extremely difficult to read,
mixing a lot of logic with a lot of formatting. Furthermore, the same
logic was duplicated between files, which could lead to consistency
issues. The new code moves all the logic to python helpers that
create data structures, the template only focus on formatting and
rendering.

Since the dt uses more complex structures than topgen, e.g. nested
structures with nested array mapping, a complete set of new classes
to deal with this complexity.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>