[rom_ctrl,doc] Update docs for drop of second S&P layer
This second S&P layer was removed in commit 803d9aec77f. Update the
documentation to describe the design properly.

While updating the block diagram, I also sort out the legends for the
rdata paths.

In the programmer's guide, I change the text so that it's
true (talking about scrambled addresses as opposed to scrambled data),
but the behaviour seen by the programmer does not change.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>