[tl,dv] Clear tl_agent_cfg::a_source_pend_q when leaving reset
Before this commit, the tl_host_driver asserted that the queue was
emptied during reset. Items normally get removed from the queue with
tl_agent_cfg::remove_from_a_source_pend_q function.

This would normally be done by tl_host_base_seq after it got a
response from a TL transaction. Practically speaking, this will also
happen when there is a reset, because the driver sends an instant
dummy response on reset, but it probably makes sense to be really sure
the queue is empty by the time the reset completes.

If it isn't empty for some reason, there's no real point in failing
the test, so this commit changes the behaviour to clearing it. The
commit also moves the operation to tl_monitor, which is probably a
better place to update the cfg object.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>