[tl,dv] Document, rename and fix chk_prot_mask_in_enabled_lanes
This behaviour actually changes slightly because we base the
calculation of active lanes on DataWidth instead of SizeWidth. The
latter is TL_SZW.

For a 32-bit bus, TL_DW is 32. As such, byte lane zero is all
addresses that are multiples of 32/8 = 4.

In the existing code:

 a_addr[SizeWidth-1:0]    is the index of the lowest active lane.
 (a_mask >> ..)           is the mask, shifted to put the lowest
                          active lane to the start.
 (.. >> (1 << a_size))    is the mask shifted by another 2^a_size bits.

This checks that none of the lanes above the active lanes have a
nonzero mask.

In the new code:

 (1 << a_size) - 1            is a pattern of the bits for active
                              lanes, starting at index zero.
 .. << a_addr[SizeWidth-1:0]  is the pattern of active lanes, starting
                              in the right place.
 ~(...)                       is the pattern of nonactive lanes
 a_mask & ...                 is the pattern of inactive lanes that
                              are HIGH in the mask.

This should behave the same for the lanes above the active lanes but
should also require the mask to be zero for lanes below the active lanes.

The other change is a minor tweak, and changes discussion of "enabled"
lanes to talk about "active" lanes (matching the terminology in the
spec).

Change-Id: I8f2047f1398952db53bb5d730cfbd82385b9b119
Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>