[fusesoc] Move all cores under the hw directory
Avoid conflicts with various build output directories that get placed at
the root of the repo, including bazel's symlinks, by placing fusesoc's
cores-root in a subdirectory (specifically, the hw/ subdirectory).

Move the check_tool_requirements core and scripts, plus the
tool_requirements.py script, to the hw directory, so there is just the
one cores-root for the opentitan repo.

Signed-off-by: Alexander Williams <awill@opentitan.org>