[rom_ext] fix CDI_* update bug
If non-CWT CDI certs are found in flash info page 19, we should treat
this case as if the cert page is empty, and regenerate the certs. This
enables running an X.509 ROM_EXT first and changing to a CWT ROM_EXT
later. Since UDS cert is never generated outside perso (and is on a
separate flash page), the UDS cert format is defined at provisioning
time.

Change-Id: I0d8b6b8bfb775fdb9db06e46d18568b97b1add9f
Signed-off-by: Tim Trippel <ttrippel@google.com>