[gpio] Add input period counter feature
Each of the input period counters can count the number of clock cycles
in one period (i.e., from one rising or falling edge to the same edge in
the same direction) of an input signal.  Each input signal can be
assigned to any of the counters.  The counter can operate in one-shot
(i.e., just measure one period) or continuous (i.e., keep measuring
periods) mode.  The polarity (i.e., whether a rising or a falling edge
counts as begin and end of a period) is configurable.  A prescaler
extends the range of the period counter.

Change-Id: Idf69412ed359903c7e3209814779167dd1555e05
Signed-off-by: Andreas Kurth <adk@lowrisc.org>