[lc_ctrl] Buffer hardware revision to allow differentiating metal spins
This commit places a prim_sec_anchor_buf on the constant hardware
revision signal inside lc_ctrl. This prevents synthesis from performing
logic optimizations on the hardware revision CSRs and output signals and
ensures the individual bits of the constant remain visible in the
netlist. This enables differentiating metal spins of the same design via
the hardware revision in the future.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>