[sw] Port hmac_enc_test to devicetables
The `isr_testutils` API is not yet ported to DT, and this test only uses
a small subset of the functionality exposed by `isr_testutils_hmac_isr`.
So, for this test, we manually replace this function call with a few
lines of equivalent dt code to keep the test passing. Since we no
longer use `isr_testutils`, this drops the corresponding `isr_ctx` and
`hmac_ctx` structs in favour of simplifying the test implementation.

Change-Id: Ic3c5c17dc87e61ade8debc16ffe1034bdffdcf6c
Signed-off-by: Alex Jones <alex.jones@lowrisc.org>