[bazel] Fix `opentitan_select_top` documentation
The docs for this function use non-existent top `english_breakfast` as
an example (`englishbreakfast` is the actual name). They also give an
example of using an array as a dictionary key, but arrays are
non-hashable and so a tuple must be used instead to make a collection
as a key.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>