[sw] Port `spi_passthrough_test` to devicetables
This test requires more significant changes to be suited to running on
Darjeeling, as it uses some Earlgrey-specific pinmux mappings, and uses
two SPI Hosts instead of just one. Only the first SPI Host is actually
used, so we enable the test to scale to initialise any number of SPI
hosts per top but still only use the first. Pin mappings are hard-coded
for Earlgrey/Darjeeling based on the top-level using conditional
compilation.

When porting to DT, more generic logic for enabling all the interrupts
for the Spi Host and Spi Device blocks is introduced, to avoid
hard-coding the name of interrupts when the test itself does not require
it.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>