[top_darjeeling] Build fixes for remaining DV sequences
Build fixes for the remaining top-level DV sequences for
Darjeeling. This does _not_ mean that they function correctly
yet.
Most changes are related to otp_ctrl becoming an ipgen block.
The remainder are due to the absence of the sensor_ctrl and
thus 'ast_init_done' status.
Fix a few warnings from the build log to keep it cleaner.

Change-Id: I3c6d1dbace4d489be88dcc8e72c2119c97145efe
Signed-off-by: Adrian Lees <a.lees@lowrisc.org>