[opentitantool] Make HyperDebug firmware upgrading more robust
Updating firmware involves resetting HyperDebug to jump into a DFU
bootloader, and back to the main firmware.  Previously, we changed
opentitantool to allow more than one second for the jump from DFU to
main firmware, to allow for various cases in which the USB enumeration
could take longer.

Recently, I noticed a case where one second was not enough for jumping
from the old firmware to DFU bootloader.  It happened that another USB
device misbehaving on the bus seemed to cause the Linux kernel to take
longer to get around to recognize the "new" DFU device on the bus.
(This misbehaving USB device was OpenTitan, which had abruptbly been put
under reset, as HyperDebug prepared to update firmware, and stopped
driving the OpenTitan POR signal, leaving it with a 1Mohm pulldown.)

This patch applies the same retry-for-5-seconds logic when doing the
initial jump to DFU, as we currently have for jump out of DFU.

Change-Id: Ia1c61f4a74c5132bfde7e0125c48dd88c9f0e89c
Signed-off-by: Jes B. Klinke <jbk@chromium.org>