[dtgen] Expose the alert count in the DT API.
The `all_escalation_resets_test` needs to know the number of alerts as a
compile-time value constant so that it can reason in a generic way over
the alerts. Expose this as an enum constant in the global DT API along
with other alert information.

Change-Id: Id705c28723ef3846b1067eb0bbf1695663c6d7a3
Signed-off-by: Alex Jones <alex.jones@lowrisc.org>
(cherry picked from commit ffd84cd4ce53fc61c49763bb416a6d3d48697a51)