[sw,dma] dma_inline_hashing bring up on Darjeeling
Changes to make (chip_sw_)dma_inline_hashing functional
on Darjeeling. These involve the modified addressing modes
of the DMA block (using 'wrap' and 'increment' configuration
parameters for each of source and destination).

Additionally the error interrupt is enabled and detected in
case the test fails. This is just for diagnostic use and permits
a test failure to complete more quickly.

Note that presently the OT build system does not map the
'.ctn_data' section into the CTN address space so this cannot be
used successfully as a destination without further work.

Change-Id: I29f0493cef53c6eb7df5fbd811a162b8ae5dd159
Signed-off-by: Adrian Lees <a.lees@lowrisc.org>