[ottf] give `test_status_set` a release semantics
For DV tests, a frequency flow is:
1. Device SW do something
2. Device SW set a test status, and DV vseq detects the change
3. DV do something
4. DV writes a backdoor variable, and device SW detects the change
5. Device SW do other work

For this to work properly, we need 1, 3, 5 not to be reordered.
For 3, 5 reordering, we need an acquire semantics, which is done by
OTTF_BACKDOOR_READ. For 1, 3 reordering, we need the SW -> DV
info exchange to have a release semantics.

Now, it's quite unlikely that compiler would do such reordering, and
Ibex is sequential so it won't do something strange itself. But to
ensure correctness, it's still a good practice to mark synchronization
as such and not just hoping that code compiles correctly.

Change-Id: Id4606afd4f52a7a689f1a41ce37b801ef8d9ce4b
Signed-off-by: Gary Guo <gary.guo@lowrisc.org>