[testutils,pinmux] Introduce new function to connect pad/IO
This higher-level function provides a convenient way to connect a
peripheral I/O to a pad without needing to know if it's an MIO/DIO:
it will do the natural thing in all cases. This is useful for multitop
because Earlgrey and Darjeeling have different kinds of pads, e.g.
the GPIOs are MIOs on EG but DIOs on DJ.

Change-Id: I55567cfecca0a2842d43dcc7c0b19f780505c02c
Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>