[rv_plic,rtl] Avoid an unnecessary ternary in a loop
This behaviour is unchanged, but I think it's rather easier to read.
The idea is that "set[i]" should be true if src_q[i] is true unless
le_i[i] is true (so we're edge triggered) and src_q[i] is true (so
le_i[i] was already true last cycle).

Change-Id: Ib9b921822a0f566ab8d9956ebb25acaf8efabeef
Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>