[prim, rtl] Deadcode else statement when state_q is IsStd
diff_check_ok true means that we have differential pairs 2 cycles
ago (with some amount of random cdc delay). We further flop the
output of the 2sync flop to diff_pq and diff_nq. If diff_check_ok
is false, this means the the pairs are no longer differential
atleast 2 cycles before. But since we flop the synchronized output
to diff_xq registers, means that the synchronized output pairs on
last cycle were differential.

With the above explanation, in state IsStd when diff_check_ok is
no longer true, we look for diff_n_edge and diff_p_edge and they
must be differential. This leads state_d to jump into IsSkewed.
So, this is impossible to jump into SigInt when the state is
IsStd.

The reason why we care about this change is jasper analyzed the
coverage of this else block as dead. This PR tweak the case for
state is IsStd, to just look for diff_check_ok and it is false
then jump into IsSkewed.

This change has also got checked through an assertion saying
if state_q is in IsStd then we can either have diff_check_ok or
diff_p_edge and diff_n_edge are differential.

Change-Id: Ifa6fcd9754e26bcb010b4b5192ac110f7491ce8a
Signed-off-by: Kinza Qamar <kqzaman@lowrisc.org>