[dif,rstmgr] Add rstmgr DIF for getting SW reset index from DT reset
For space efficiency the rstmgr DT extension only exposes this
information in one direction, mapping from SW reset indexes to DT
resets. This test requires the ability to map from given devices/resets
to the corresponding SW reset index. Implement a DIF function with a
basic linear function for this purpose. Due to the small number of
software resets, this shouldn't introduce a large performance overhead.

Change-Id: I08d97fa2b6e389f7f3774c40d46ba8ba82dfb6d0
Signed-off-by: Alex Jones <alex.jones@lowrisc.org>