[dv] Constant-initialise a dynamic array more nicely
Technically, the version that we were using didn't actually allocate
any memory for the array. Practically speaking, it seems that VCS and
Xcelium both do what we want, but Verissimo points out that it's not
technically standards-compliant.

Fix it.

The fix changes LIST_OF_ALERTS to no longer be a dynamic array. It
turns out that there is *one* place in the repository that
cares (chip_sw_alert_handler_entropy_vseq.sv), so this commit
expresses that code very slightly differently.

Change-Id: Iefa3691bce9546f89c26b59a3b2834a29296098b
Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>