[dv] Simplify use of dv_base_reg::atomic_en_shadow_wr
This code was trying to ensure that the semaphore had exactly one
token in it. Knowing that it would only ever have at most one token,
the code worked by trying to get that token if there was one and then
putting a new token in.

Casting the result of the try_get() function to void causes a lint
error from Verissimo. And there's a simpler way to get the same
behaviour: just make a new semaphore!

This would be a problem if anything took a reference to the old
semaphore (leaving a "dangling pointer") but a quick grep shows that
this doesn't happen.

Change-Id: I686a2df9db4314380b72300a40dcdd8fe37092a0
Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>