[rules] Make opentitan_test expose all files in DefaultInfo
There is a discrepancy between opentitan_binary and opentitan_test.
The former exposes all files which were built, while the latter only
exposes files in a special list (PROVIDER_FIELDS). This is quite
inconvenient when we want to add new fields to providers. This PR
removes the special list and instead makes opentitan_test expose
all files its in provider.

Note that this change means that the default info will now
potentially be exposing fields which were previouly only available
in the provider but this should not cause any issue.

Change-Id: I604c52668fc7ac90a69cffff021418922b033961
Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>