[racl_ctrl,dv] Template racl_ctrl_sva
The racl_ctrl_sva core uses the templated hjson to work out what to
do. Pull that into a template (which can then find the relevant
hjson).

This commit also does the same to racl_ctrl_bind, which is a little
silly because that doesn't really need to be templated. But it's part
of the same core and not very long.

Change-Id: Idde8000e6ec11284a51812c8afc6bd89470e26d7
Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>