[usbdev] DPI model spends a little over 1ms in Bus Reset
The DPI model spends slightly over one bus frame in the Bus Reset
state after detecting a device connection. This is now 1ms since
the duration of the bus frame was originally shortened
artificially to reduce simulation times. It was subsequently
extended to model the real USB.

With a 1ms timeout the test has been observed to fail sometimes
because the software gives up just as the transmission of the
first SOF packet commences. Increase the timeout to 1.5 USB bus
frames.

Change-Id: I1b5f931b22b063c2643d7d2cf7adbd7a9c3e11b1
Signed-off-by: Adrian Lees <a.lees@lowrisc.org>