[ci] Allow `+` in Bazel target names
This has come up a couple of times recently with names containing `C++`
which are allowed by Bazel but banned by this rule.

I don't think a `+` could cause the kind of type confusion that the
comment here talks about. Also Bazel 8 uses `+` in canonical target
names so we may need to accept them anyway if `query` ever spits out
one of those.

Change-Id: I12d2f5e7f090474de22b4993bc45b3ab4deebe96
Signed-off-by: James Wainwright <james.wainwright@lowrisc.org>