[fpga,rom] Speed up ROM splicing by omitting contiguous addresses
We take a reasonable assumption from the UpdateMEM documentation and
intuiton that UpdateMEM's runtime scales in some way with the number of
addresses (i.e. "non-contiguous writes", even if the data is contiguous)
in the ROM to splice.

Currently, when generating an UpdateMEM-compatible MEM file for the ROM
(unlike the OTP), we omit an address for each individual ROM word. This
is despite the fact that, for the majority of widths, the number of
characters per word will be such that data is written to contiguous
addresses.

Introduce logic to check whether ROM word addresses are contiguous, and,
if so, omit the redundant address from the generated MEM file. UpdateMEM
will then produce identical bitstreams with the new ROM file, but basic
empirical testing suggests that ROM splicing speeds up by up to 4-5x.

Change-Id: Ib3d6b0726a764cac5ef6dc69d71b0d38473944ac
Signed-off-by: Alex Jones <alex.jones@lowrisc.org>