[prim] Add single-cycle variant of `prim_ram_1p_scr`
Ibex (data and tags of instruction cache) and OTBN (data and instruction
memory) use `prim_ram_1p_scr` in a particular way:  They ignore the
`gnt_o` (and, in the case of Ibex, also the `rvalid_o`) signal, hence
assuming that `prim_ram_1p_scr` is always able to handle requests and
will provide a response in the next cycle.  This is true for
the current implementation of `prim_ram_1p_scr`, but it means `gnt_o`
and `rvalid_o` cannot be used for flow control, which is problematic for
upcoming changes to `prim_ram_1p_scr`.

This commit thus creates a variant of `prim_ram_1p_scr` with a `_1cyc`
suffix.  This variant is always ready for requests and always responds
in the next cycle.

As there's a circular dependency on this change with the Ibex repo,
which gets vendored into this repo, this commit applies a patch to Ibex.
Once this commit has been merged and the Ibex repo has been updated,
that patch can be removed.

Change-Id: Ie0ea008ba5d35b0d8911368c1491dc837f5d1196
Signed-off-by: Andreas Kurth <adk@lowrisc.org>