[rv_plic, dv] Fatal Alert never gets deasserted unless reset happens
There is an else block in rv_plic_reg_top which raises err_q high if
intg_err or reg_we_err happens. This else block was analyzed as
unchecked by jasper. To make it marked checked, an assertion is added
which means that unless reset happens, fatal alert remains high
forever. The err_q assigned to intg_err_o which is an alert for
alert_sender in rv_plic. So, the added assertion solves the problem
related to the else block unchecked coverage.

Change-Id: I088ba159a7a668f533fd7c3112a38182f8c4930f
Signed-off-by: Kinza Qamar <kqzaman@lowrisc.org>