[imm_section] Add version string to the compiled binary
This commit stores a version string in the compiled immutable section, and
prints it at the mutable stage to make it easier to debug issues between
imm_section releases.

Since the imm_section does not have a manifest, and the entry point must at
the beginning, the version is stored at the last 8 bytes of the immutable
section as a null-terminated printable string.

Change-Id: I6c669349ab6700cf91e75355fb1f80328c2507f5
Signed-off-by: Yi-Hsuan Deng <yhdeng@google.com>
(cherry picked from commit 5abc0319fb453b05a5452a0eefde8b6762d08d79)