[fusesoc,pwrmgr] Move pwr_lc_req/rsp struct to lc_ctrl
The pwr_lc_req/rsp struct is used by both pwrmgr and lc_ctrl. It is
top-independent, so it is best to place it in lc_ctrl, since it is a
generic IP, and placing it in pwrmgr causes lc_ctrl to depend on
top-specific packages.

Remove some unnecessary virtual core declarations.

Fixes #26755

Change-Id: Ibd64e0ade368677f967a050c00d53034f062f4f4
Signed-off-by: Guillermo Maturana <maturana@opentitan.org>