[sram_ctrl] Update sram_ctrl_smoketest for Darjeeling using DT
Port the sram_ctrl_smoketest to use DT and perform bring up on
Darjeeling.

Generalize the code to support a variable number of SRAM Controller
instances. kDtSramCtrlMain still needs to be treated as special
to avoid corrupting the main SRAM content.

Overlap the testing of the individual SRAM Controller instances
to increase the chance of catching an interactions/crosstalk.

Change-Id: I3b8f3d1525e19a0e6a0fb77d0dbadb1dd4b181ba
Signed-off-by: Adrian Lees <a.lees@lowrisc.org>