[darjeeling] Fix undriven input ports
There was no driver on otp_cfg, although since that's a DFT input it
has not led to simulation failures.

The only other chip-level Z/X signals are all a product of DIO/MIO
signals being undriven and are harmless provided that the pinmux
(for MIOs) or output enables (DIOs) are set appropriately before
the signals are used.

Specifically, MIOs (excepting 0 which is the TAP strap input) are
left floating (Z) when unused, and the SPI interfaces are undriven
in the cast of DUT SPI Host 0 output and connected to X in the
case of DUT SPI Device 0 input.

Change-Id: I8cb348520c29759d176294bc8d8a4abea2c6bb71
Signed-off-by: Adrian Lees <a.lees@lowrisc.org>