[tl,dv] Use capitals for macro names
This will follow the style guide more closely. (Even though the macros
get undefined again, there's no reason to have confusing names).

Also, put the undef lines in the other order so that the definition
scopes nest nicely. This clearly doesn't matter, but we may as well be
super consistent.

Change-Id: I444fd4c8f5cbb8ce2a5f3758475d33689f093a30
Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>