[darjeeling] Update config files for some overlooked tests
Modify the sw test image location for some tests that have been
ported to DT but have not been operational on Darjeeling as a
result of this oversight in the configuration file.

Some tests are no longer dvsim-only targets, so update their
filenames.

Darjeeling does not have a sensor_ctrl module, so remove those
tests which still make reference to it.

Change-Id: I3f163789037a0dd18d44b701fd15ffa59f51574e
Signed-off-by: Adrian Lees <a.lees@lowrisc.org>