[mdbook] Add patch for .mdbookignore
Mdbook has no way to ignore files when copying the website. There
is an unmerged PR (https://github.com/rust-lang/mdBook/pull/1908)
which adds a .mdbookignore file to do that. This commit cherry
picks this PR as a patch. This is not entirely straightforward
however:
- the PR modifies files which are not found in the release source
  (like the documentation)
- the PR modifies Cargo.toml but the git's version differs from
  the crates.io version (which normalizes Cargo.toml)
- the PR applies on the latest version, but we use version 0.4.43

For this reason, the patch is a trimmed down and manually edit
version of the PR. Furthermore, there is another difficulty:
the patch makes `ignore` a required dependency when previously
mdbook only had it as an optional (and not enabled) dependency,
it also bumps the version. For this reason, this commit manually
adds the required version of `ignore` to the Cargo.toml and uses
rules_rust crate annotations to add a dependency from mdbook to
this version of `ignore`.

Change-Id: I5efc8cc662848d2fb60d2d5aaecd700b4f099109
Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>