[dv, rom_ctrl] Wait for ping ahead u_checker_fsm.u_compare.start_i error inject
Before injecting errors in `start_checker_q` ensure there's not an ongoing ping.
This is done first by checking the VIF hasn't already detected a ping, and then
waiting for the ping to finish.

The first VIF ping state check is neccesary, because the alert_monitor sets the
`active_ping` flag with a 2-cycle delay, and the ping may be missed in some
scenarios depending on timing otherwise.

Change-Id: I38a2f1755b550227976f7a1878e5e77cc67d6ea8
Signed-off-by: Antonio Martinez Zambrana <amz@lowrisc.org>