[rom_ctrl,dv] Tweak "LocalEscalation" in corrupt_sig_fatal_chk_vseq
The existing code was probably doing something reasonable, but the
explanations were rather confused (claiming that the FSM is linear
when it isn't!)

Rewrite things with lots of comments. I suspect the behaviour will be
exactly the same, but this should be slightly easier to understand.

Also, avoid the last dut_init call (since we'll do another one when we
drop out of the case statement).

Change-Id: Id57f6300fe78687195b00107b634bfde6a0bcbde
Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>