[rom_ctrl,dv] Tidy up CompareCtrlFlowConsistency check
This wasn't all that clearly thought out, so I've added some detailed
comments about what we're doing (and why). With the existing code,
it *might* have started the checker a second time after it was done.

That presumably didn't happen very often (because the test would have
failed) but we don't really care about that case: the attacker is just
forcing the hardware to confirm what it already told us!

Change-Id: Ib9a4a9d48c6fd81b2b17c0511f3523119cff1fc1
Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>