[rom_ctrl,dv] Tidy up the way we try to corrupt the compare module
I'm not really convinced that the original author knew what we were
stimulating and testing. And the "CompareCtrConsistencyDone" branch
didn't really check the right thing.

Change things so that they make a bit more sense and add lots of
comments to explain what is going on.

Since we are touching the way that signals in the counter module get
forced, switch to using the new rom_ctrl_compare_if interface (which
lets us get rid of some uvm_hdl_* calls).

Change-Id: I84e27744ea1d30333bea8e758621e5cc80bd63ac
Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>