[opentitanlib] add GPIO indicator pin to SpiConsoleDevice
Currently the SpiConsoleDevice polls the DUT's SPI device buffer for
data that it wants to transmit out of the device to the host. To support
less noisy, non-polling environments, this adds an indicator GPIO pin
that can alternatively be sampled to indicate when data is ready to be
transmitted out of the device.

Change-Id: I6cf38383e1d491c54cf39f870660b7390d78c642
Signed-off-by: Tim Trippel <ttrippel@google.com>