[test,flash_ctrl_idle_low_power_test] Correct check for sleep abort
The test expects the sleep to abort because of the inflight flash
transaction. Currently, this is done by calling looking at the reset
reason stored in retentation by the ROM. This is incorrect since
if the device wrongly enter (shallow) low-power sleep then it would
not reset so the ROM will not run and rstmgr will play not role.
Instead, the correct way to check this is to ask the pwrmgr's wakeup
info and check for the aborted type.

Note that the test by itself was still performing as expected because
the aon_timer is configured in a way that if the device ever entered
sleep then it would reset due to an watchdog bite and this would
cause the test to fail anyway. But it would not catch other types
of errors such as the pwrmgr entering sleep and waking up unexpectedly.

Change-Id: If5f122a4b06c7c562c3a2f87037f95e8513de641
Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>