[dv,flash_ctrl] Fix flash_op randomization of partition
Avoid illegal partition values in rand_op.partition constraint in
flash_ctrl_otf_base_vseq.sv. This was causing random failures that
didn't match legal partition values. The particular setting causing
problems is not really representable in the hardware settings, so it
is not even a good test for invalid ops.
Double the randomization probability of data partition accesses.

Change-Id: I2d46cc4ee7bc75cd7e3bd1efb09ce310169a25ab
Signed-off-by: Guillermo Maturana <maturana@opentitan.org>