[dv] Iterate more carefully with foreach loops
The changes in PR #26908 don't quite do what we want. Before that PR,
we had code like

  foreach (foo[a]) {
    foreach (foo[a][b]) {
      ...
    }
  }

Some lint tools warn that the inner foreach loop is badly formed, and
I think they might be correct. The IEEE 1800 spec doesn't seem to
allow this syntax. The start of section 12.7.3 (about the foreach
loop) says:

  The foreach-loop construct specifies iteration over the elements of
  an array. Its argument is an identifier that designates any type of
  array followed by a comma-separated list of loop variables enclosed
  in square brackets.

This doesn't match the syntax for the inner loop in the example
because "foo[a]" is not an identifier.

The changes in the PR converted that example to something like

  foreach (foo[a]) {
    foreach (foo[a, b]) {
      ...
    }
  }

This is definitely reasonable syntax but Matute correctly points out
that it doesn't do the same thing as VCS used to do with the previous
code. Now we iterate over "a" twice!

This commit does what I think is probably the correct fix, dropping
the outer loop in this example.

There are a few examples where the change isn't completely trivial.
These previously looked like

  foreach (foo[a, b]) {
    foo[a][b].size() = ...;
    foreach (foo[a][b][c]) {
      foo[a][b][c] = ...;
    }
  }

Here the inner loop really does want to iterate over c for some fixed
a, b. This commit changes these to look like

  foreach (foo[a, b]) {
    foo[a][b].size() = ...;
  }
  foreach (foo[a, b, c]) {
    foo[a][b][c] = ...;
  }

The previous code was trying to say "For each array, constrain its
size and then iterate over the items of the array, constraining their
values." Now it says says "For each array, constrain its size. Now
constrain the value of every item of every array."

Equivalent, and avoids the nesting!

Another rather fiddly example is flash_ctrl_mp_regions_vseq. Here, we
have an array called mp_info_pages where the inner-most array is
dynamic. Annoyingly, fixing i and then iterating over e.g.
mp_info_pages[i][0][k] isn't really possible. A cheeky solution is to
actually iterate over all the indices with mp_info_pages[i,j,k] and
then condition the constraint with j == 0.

Change-Id: Iff1507fb09b7124d53287596a5c3e847f9e4c373
Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>