[dv] Avoid bare "wait fork" in cip_base_vseq
The code that was in check_no_fatal_alerts had a bug because it used
wait fork without an isolation fork, meaning that we might end up
waiting for any other processes that had been started by other bits of
code.

The Verissimo lint tool generates an error to warn about the mistake
and this commit tidies things up a bit.

To avoid the code structure getting overly complicated, the commit
also splits out the "wait for a bit, waiting for an alert" code from
the iteration over different alert names.

It also extends some comments a bit, in particular the comments that
explain how we wait for any alert's ack to have been cleared (figuring
that out took a certain amount of git archaeology!)

Change-Id: I28124ad3d0ec1fa5bec241e205859d786e45b642
Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>