[dv,flash_ctrl] Fix misuse of reset states
The DV code uses its own enum for the states in flash_ctrl_lcmgr, and
flash_ctrl_hw_rma_reset_vseq needs to do it as well. The code was instead
comparing the dv enum to the hardware states which is a type mismatch
and makes no sense, since the hardware enum is wider for redundant
encoding.

Change-Id: I89427e42215b79380ecc3fc9c165a2948bc53cbb
Signed-off-by: Guillermo Maturana <maturana@opentitan.org>