[topgen,otp_flash_keys] Deduce if flash_keys are needed in otp_ctrl
The otp_ctrl interface to export the flash_keys is only needed if
the otp secret1 partition holds flash_ctrl keys. Darjeeling has no flash
and has no keys in otp.

Detect this examining the content of the otp rather than using an
ipgen_param.  This assumes the flash keys are in the secret1 partition,
and are named "flash*key_seed" (case doesn't matter). If in some future
otp mmap the location of these keys changes we can revisit this
detection.

Addresses #26553
Fixes #26931

Change-Id: Ib68f5f78838361327f5b6626a4447be8b0f2cef1
Signed-off-by: Guillermo Maturana <maturana@opentitan.org>