[rom_ctrl,fpv] Define two helper properties
The StabilityChkkeymgr_A target property was not converging. You can
generate a 3-cycle SST example with

    prove -property tb.dut.StabilityChkkeymgr_A -sst 3

and the result shows something where `in_state_done` is true in the
FSM, but `checker_good` changes. The root problem seems to be that the
k-induction search doesn't know that the compare module thinks it is
done when the surrounding checker FSM does.

The LastImpliesValid_A assertion needed debugging in a similar way.

Co-authored-by: Marno van der Maas <mvdmaas+git@lowrisc.org>
Change-Id: Iedb5abc8240f22edcbdc7dbd5d745475035be5ac
Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>