[rtl,sec] Define more precise version of ASSERT_ERROR_TRIGGER_ALERT
Existing properties like ASSERT_PRIM_COUNT_ERROR_TRIGGER_ALERT aren't
actually true in a formal setting! The problem is that they measure
the time between a bad thing happening and the alert_p signal going
high. But there is a stateful interface between the alert sender and
receiver, so that time depends on the behaviour of the alert receiver.

I started thinking about how to reason carefully about this, but
realised (with help...) that it was silly! We actually have formal
verification of the alert machinery (for example, see the
lowrisc:fpv:prim_alert_rxtx_fpv core). As such, it probably makes more
sense to be checking that after the bad thing happens, the error is
passed to the alert system. Then the existing (shared) DV for the
alert system will check that the alert_tx_o signal is updated
accordingly.

Change-Id: Iaf0e3ba77c55152da5863cb627ed52afc483f00c
Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>