[otlib] Improve serial port RX buffer clearing
When clearing the uart RX buffers, there are two buffers to take
care of: the host buffers and the device buffers. The host buffers
are taken care of, but in the case of CW310 (which uses the SAM3x),
the code is not currently doing anything for the device buffers.
This is different on other backands such as hyperdebug which have
a dedicated usb request to achieve.

This commit changes the behaviour of the default "serial port uart"
implementation to both clear host buffers and try to read from the
device until we get no more data (or a timeout expires). This change
will primarily affect the cw310 (sam3x) backend. On hyperdebug, since
we first clear the device buffers and then the host buffers, reading
from the device will not yield any data and this change will only
result in a short (5ms) delay.

Change-Id: I4203a6cbe391b7798047bfdaf22864b4e6a8bea1
Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>