[ottf_console] optimize TX-ready GPIO feature
This optimizes the TX-ready GPIO feature by only writing one frame at a
time to the SPI flash buffer and monitoring the chip-select state for
when the device is free to overwrite the data in the buffer (i.e., the
host has read out the data). This eliminates the need to add delays
between TX-ready GPIO toggling.

Change-Id: I320b80154807aa930ab99a6846101f2cfde352e6
Signed-off-by: Tim Trippel <ttrippel@google.com>