[multitop] Move top description out of generated files
The top description (created was `opentitan_top`) was previously
created in `<top>/data/autogen/defs.bzl` which is inconvenient
because it contains attributes that one may want to manually specify.
This commit moves the call to manually created file at the root of each
top's directory. The list of IPs for each top is still created by topgen.

Change-Id: I15068814aad0f990fc8649c45c62c906ff924925
Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>