[rom_ctrl, dv] tl_errors_cg coverpoint on write with instr type true
When we inject a write_w_instr_type_err, the coverpoint doesn't get sampled.
To get that sampled, the check in cip_base_scoreboard requires all the
errors to be onehot. But when we get write_w_instr_type_err, we also get
mem_ro_err which makes the $onehot check false and the autobin[1] never gets
sampled.

The PR removes the write_w_instr_type_err term from the check and replaces
$onehot with $onehot0.

Change-Id: I6d3f62c7c2754fb43009fbc8baa4d3d74a68189d
Signed-off-by: Kinza Qamar <kqzaman@lowrisc.org>