[aes,dv] Reformat aes_wake_up_vseq
There is no functional change here. The commit was motivated by a
Verible check (because of long lines), but I'm also rejigging things
slightly to use out-of-block definitions and avoid unnecessary class
variables.

Finally, it also drops an extra randomisation of the sequence that
happens just after we start the body task (huh?!).

Change-Id: I60ef14765d9b7d2f4573dcbd0fd51e9349d21338
Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>