[darjeeling] Update README.md
Prior to this commit, Darjeeling's README on `master` mentioned the
`integrated_dev` branch.  Now that Darjeeling's code lives on `master`,
we should no longer reference `integrated_dev` from there, though.

This commit removes the reference to the `integrated_dev` branch and
adds information on code generation and tooling that is also in
Earlgrey's README.  Instead of duplicating the information,
Darjeeling's README now links to the relevant information in Earlgrey.

In the future, we might want to template `README.md`, in order to render
the same information for both tops without duplicating it in the source
code.

Change-Id: I312f1fab9646aecba1d428db9d50318616bdea48
Signed-off-by: Andreas Kurth <adk@lowrisc.org>