[rv_plic, fpv] Alert Sender FSM faulty state
This PR usage a mechanism to inject fault in the state register
of alert FSM. This mechanism is to create a stopat at state_q
so that jasper drives any possible value. When it does, we
cover the stimuli coverage for a default case in
prim_alert_sender. To cover the checker coverage, an assertion
is added. This assertion and stopat is created in a separate
isolated task so that none of our proved assertions fails when
we are doing FI in FPV.

Change-Id: I173dcb429cdb27dcdabfaf018e6f99dd70579d9a
Signed-off-by: Kinza Qamar <kqzaman@lowrisc.org>