[soc_dbg_ctrl] Assign register fields individually
Assigning a packed reg2hw struct to a corresponding hw2reg struct
is fragile and at the time of writing incorrect because it is
reordering the fields.

These packed structures do not represent the register contents as
such and may be modified by changes to reggen or indeed to other
fields within the register, such as the insertion/removal of
the enable signals (e.g. 'qe' and 'de').

Change-Id: I263c873deefe92a64f951bb597fd25e16c783428
Signed-off-by: Adrian Lees <a.lees@lowrisc.org>