[crypto] Implement key integrity checksums.
Uses a CRC32 checksum to provide a lightweight integrity check on keys. The
basic setup for this already existed, this commit just:
- fills in the stub
- corrects a few places that didn't write the checksum
- changes a couple fields in the key structs from `size_t` to `uint32_t`

The last item is so that we have consistency across platforms in how large the
field is and can use `crc_add32` safely for host-side tests as well as
device-side ones.

Change-Id: I03c05de9ac071b9023b232f1f624de9fb985225f
Signed-off-by: Jade Philipoom <jadep@zerorisc.com>