[rv_plic, fpv] Waived the statements as UserInIsZero=1
The statement that are waived in this PR are appears to be
undetectable coverage. Those were the assignments to the
output tl_o.d_user.*_intg. The instantiations was having the
parameter UserInIsZero=1 which will generate
tl_o.d_user.*_intg as zero. Hence, leaving them waived is
reasonable rather than making an unreasobale assertion.

Change-Id: I5fe0ed6b9f1175326279544c4e931e7c2a2d92d6
Signed-off-by: Kinza Qamar <kqzaman@lowrisc.org>