[ottf,console] Make buffering more general and make it per-console
Currently buffering is SPI only and also uses quite a few global
variables. This commit makes buffering more general since it has
nothing (fundamentally) to do with SPI. The approach here is to
move all buffering related variables to the console object. Since
buffering requires a staging buffer, this one has to be provided
during configuration and for the main console, the global buffer
is used for that exact purpose.

Change-Id: I4ed33f97c5f6facec80ed3285d7f86f89920fd1f
Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>