[doc] Tweaks to otp_ctrl documentation
A few minor changes to fix some issues with the OTP Controller docs.
The main goal was to fix the links that got broken when the docs were
split into multiple files.
Also removed a link that didn't seem to make sense.
While making changes, also clarify some of the text around the links.

The other main change was to strip the newlines out of the partition
descriptions when building the memory map table to avoid each
partition being split in to many lines depending on how long a
description it has.

Also fixed a typo and removed a random 's'.

Change-Id: Ic2ce24cb5d37ff201f5e064a2c15ad9d58a8975a
Signed-off-by: Elliot Baptist <elliot.baptist@lowrisc.org>