[uartdpi] Exit string for simulator exit
This adds an option to the UART DPI to define an exit string that when
seen will exit the simulator. This is useful as a way for software to
exit the simulator where there are no other ways to exit the simulator.

By default this exit methodology is disabled so should be safe to
integrate into OpenTitan. It is a more general implementation for the
exit methodology already used in Sonata:
https://github.com/lowRISC/sonata-system/blob/a9142361a190da2a196556b658f516cb9abe90b6/vendor/patches/lowrisc_ip/uartdpi/0002-Simulator-Exit-Condition.patch

Change-Id: I5617a7a62f2e37ed0f8dc63a7ce4a625ff2476df
Signed-off-by: Marno van der Maas <mvdmaas+git@lowrisc.org>