[rv_dm/rtl] Allow TL-UL transactions to complete on DMI disablement
Prior to this commit, disabling the DMI (controlled by lc_ctrl) would
directly cut off handshakes between the TL-UL and DMI bus.  Any TL-UL
transaction that was outstanding when the DMI gets disabled would not
complete because it never got a response.

This commit lets responses from the DMI propagate to TL-UL if there are
pending responses when the DMI gets disabled.

Change-Id: Ia20c7382ee7df81e5d39e403f08bab0a40dca060
Signed-off-by: Andreas Kurth <adk@lowrisc.org>