[flash_ctrl/dv] Fix path used in FI tests
The path previously referred to a generate case in which the FIFO would
have more than one entry, but `u_rspfifo` has as many entries as the
value of the `Outstanding` parameter of `tlul_adapter_sram`, which is
not set for the `u_to_rd_fifo` instance in `flash_ctrl`, hence the
default is 1 entry.  This caused the `flash_ctrl_sec_cm` test to fail.

This commit fixes the path to be correct for the singleton FIFO.  The
`flash_ctrl_sec_cm` test passes again.

Change-Id: I16ea1baeefab343e6a464aaae62a4a75e041d973
Signed-off-by: Andreas Kurth <adk@lowrisc.org>