[opentitanlib, testutils] Add UART bitbanging encoder interface
Adds an interface for encoding UART transmisions (writes) over the TX
pin as bitbanging samples, to be used with e.g. the gpio_bitbanging
transport interface. This functionality is written in a self-contained
way (only dependent on `anyhow` and `serialport::Parity`), and so can
easily be used outside of opentitanlib if needed.

Change-Id: Ia0d9f91dcb9db869ce235955a2958029a38ed070
Signed-off-by: Alex Jones <alex.jones@lowrisc.org>