[opentitanlib, testutils] Add UART bitbanging decoder interface
Adds a utility for decoding UART transmissions from a series of
monitored GPIO samples (in the form of a uniformly sampled waveform,
sampled at the correct baud rate). As before this is kept isolated from
opentitanlib so as to allow easy re-use elsewhere.

When decoding we consider a break condition to be any logical low
transmission of equal length to or longer than a UART frame (start bit +
data bits + optional parity + stop bits). We represent broken decoding
states with an output error that records the parsed data & parity bit
for reference.

Change-Id: I765b4dacf0dd9ec61d3e1ecf5d3996391656b780
Signed-off-by: Alex Jones <alex.jones@lowrisc.org>