[RIVOS][util] Add tool to generate pyreg/pydump files
This adds a new tool based on the reggen library that generates
pyreg/pydump files from IP HJSON files.

Goals for this tool include replacing most of ottopyreg, and generate
pydump files for OT internal IPs.

The motivation for having this tool is to provide a more accurate
conversion from HJSON to pyreg/pydump by using the OT libs directly, and
ease up the burden of maintenance related to the HJSON parsing code in
ottopyreg.

Related to SW-9146

Change-Id: Iebcad5655608bf0b21e67cf21f09db8d2f9adf45
Signed-off-by: Alexandre Sollier <asollier@rivosinc.com>