[usbdev] Conditional update of rxenable_out register
Avoid a sw-hw race in which the USB device may disable OUT packet
reception on an endpoint - because of the set_nak_out functionality -
and software then re-enables OUT packet on that endpoint when it
was simply trying to re-enable OUT packet reception on another
active endpoint.

This is achieved by including 'preserve' bits in the upper half of
word writes to the rxenable_out register; software may then update
only the endpoint(s) it wants to affect.

The use of 'preserve' bits, along with the priority of software writes
over hardware clears achieves backwards compatibility whilst allowing
firmware to be simplified on future versions of the USB device.

Change-Id: Ie1a5cc7233a8cda2dcdfdc5967c5e9f3b6974cce
Signed-off-by: Adrian Lees <a.lees@lowrisc.org>