[dvsim,hjson] Remove all dead code 'design_level' references/overrides
As of the completed migration to ipgen, we no longer make use of the
'--flag=fileset_X' selectors to switch out filesets between ip-level and
top-level. (e.g. `fileset_top`)
Hence, this code is now dead, and can be removed.

Note. that there are example uses of 'fileset_partner' as a suggestion for how
integrators can use this mechanism to swap out filesets. There is nothing
stopping downstream users from continuing to swap-out core filesets using this
mechanism and passing flags via the 'sv_flist_gen_flags' hjson variable.
However, the documentation should be updated to suggest the use of fusesoc
virtual cores as a preferred solution to this problem.

Change-Id: I110634ea1033803c5b373618cc84086ae27f7428
Signed-off-by: Harry Callahan <hcallahan@lowrisc.org>