[otp_macro] Name TL-UL interface `prim`
Integrators may need to customize the registers in `otp_macro`.  To do
so, they use regtool's `--alias` argument, which needs a *named* TL-UL
interface.

This commit names `otp_macro`'s TL-UL interface "prim", which is the
same name that was used before `otp_macro` was split out of `otp_ctrl`.

Co-authored-by: Andreas Kurth <adk@lowrisc.org>
Change-Id: I418814cf74eda4269139d97e1734a0362d4bea00
Signed-off-by: Eran Meisner <eran.meisner@nuvoton.com>