[alert,fpv] Fix the AlertPingOk_A assertion for async alerts
This assertion tries to check that the ping machinery works properly
in the good case: if the receiver is asked to send a ping to the
sender then it will report a response in a bounded time.

But we don't want to consider situations where skew on a differential
pair causes the single-cycle ping sent by the receiver to be
lost (lowering the skew "jumps over" the ping). Similarly, we don't
want to allow skew on the ack signal to cause a signal integrity
error. In either case, we'd end up with an alert from the
receiver (which is the safe behaviour we want).

Change-Id: I161adafae6ba49534f4af07afc679f5b61a15e07
Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>