[RIVOS][util,pyregtool] Don't require RACL info
This adds a new CLI argument to pyregtool, `--no-racl`, which tells
pyregtool that it is expected to have no information about RACL and to
proceed without it.

This prevents the generation of the `racl_group`, and `racl_ctrl`
parameters in registers of the generated pyreg/pydump file.

The `racl_ctrl_policy_name` and `racl_ctrl` parameters can still appear
on registers that have the `racl_ctrl_policy_name` tag set.

Change-Id: I71eccd648436d75b0b81203c7f32c7d28dcd17c5
Signed-off-by: Alexandre Sollier <asollier@rivosinc.com>