[opentitanlib, testutils] Fix SPI bitbang decoding `wait_cs`
Convert the sample iterator used by `wait_cs` to be peekable, and peek
until we find the first sample with CS low, instead of consuming the
first CS low sample.

If a SPI host is used with CPHA=0 and no CS hold delay, we need to
decode and read the 2nd sample (directly following CS low, where the
clock becomes active). The current logic would consume the 1st sample,
then wait for the clock to be idle on a subsequent sample (3rd sample),
and then take the 4th sample as the first bit, which is not correct.

Change-Id: I1c7306f015b53717dc17f25e8c27052780dcf599
Signed-off-by: Alex Jones <alex.jones@lowrisc.org>