[opentitanlib, testutils] Add SPI host bitbanging encoder interface
Adds an interface for encoding SPI transmissions (writes or reads) as a
SPI Host over the SCK / CS / COPI (D0) / CIPO (D1) / D2 / D3 pins as
bitbanging samples, to be used with e.g. the gpio_bitbanging transport
interface. This functionality is intended to be self-contained, so it
can easily be used outside of opentitanlib if needed.

Change-Id: I7b7f6694ca17af84293171f24ddf1d3b22ef3d8e
Signed-off-by: Alex Jones <alex.jones@lowrisc.org>