[dpi] Fix build with recent toolchain
With the toolchain I was just trying to use for a compilation (GCC
14.2.1; glibc 2.40), these two files failed to build because we didn't
have the relevant feature test macro defined.

uartdpi uses cfmakeraw, which has a complicated definition, so the
easiest approach is probably to define _DEFAULT_SOURCE.

The setlinebuf call in usb_monitor is just a specialisation of
setvbuf (which I got from the man page!). Expand it.

Change-Id: I8c466a40d07501cdb612d423a3962a9c5b2a121a
Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>