[dpi] Avoid undefined behaviour in uartdpi.c
It turns out the cfmakeraw function is supposed to be used to
customise an initialised termios struct, rather than writing the whole
thing. Do this by grabbing values from the configuration of the slave
side of the uart pty.

(The name is confusing! It should be read as "make the thing more raw"
rather than "make a raw thing")

Change-Id: Ib183bbee94d9f9ebd526f5a0793b59d555d96973
Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>