[pentest] Bugfix csr_combi test
The CSR combi test writes values to working registers to read them out later.
The triggers in between the test can interfere with those. Instead, use the
volatile register variable in C in order to make it clear that those registers
should not be overwritten to make the test stable.

Change-Id: I675603c35a7d985b0bdb79372abb057bcea75576
Signed-off-by: Siemen Dhooghe <sdhooghe@google.com>