[RIVOS][ci] Update container image build job
This updates the CI job that takes care of building and pushing the
OpenTitan container image by restricting when the image gets pushed to
the container registry:

* the image tagged with the commit hash is now only pushed when the CI
  job is run from a non-dev branch
* the image tagged latest is now only pushed when the CI job is run from
  the repository default branch

Additionally, a new image tag is pushed when the CI job is run from a
non-dev branch. It is named after the name of the branch on which the
container image was built.

Change-Id: I2b1702ea382b699f1886265c6d33afcfe70da96e
Signed-off-by: Alexandre Sollier <asollier@rivosinc.com>