[bazel,ci] Download toolchains with retry & backoff
Since we already have `urllib3` available as a Python dependency, switch
our existing HTTP GET requests to use the newer interface with support
for retries on failure with exponential-backoff. This should hopefully
work around short transient network failures and rate-limiting issues
that we occasionally encounter in our CI environment.

We need to manually implement a basic exponential-backoff scheme for the
toolchain download itself (compared to the initial query), because there
is still the possibility for mid-stream failures after a successful HTTP
status code. `urllib3` cannot handle this natively, so we wrap in
additional logic to handle this case.

Change-Id: I74623cb3b79bc0cc1e07753307d3bff2efe3f15c
Signed-off-by: Alex Jones <alex.jones@lowrisc.org>