[dv, csr] Fix prediction of non-hro REGWEN registers for hro registers
When locating the REGWEN register of hardware read-only (hro) registers,
the auto-generated DV implementation erroneously searched the list of
hro registers, instead of the list of all registers.

This bug caused DV to fail for hardware IP blocks featuring hro CSRs
controlled by non-hro registers. One example is ENTROPY_SRC for which
write accesses to several configuration CSRs is blocked through a hwo /
software read-only REGWEN CSR. As a result of this bug, the crucial
entropy_src_rng vseq had a failure rate of 100% and the functional
coverage for the ENTROPY_SRC was as low as 20%.

This is a follow-up of commit 06b911723fd485323e34749bbb7e7a741b8dbd1e .

Change-Id: I15c20cd7b296c47e99844cbf2df77042cebbc028
Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>