[dv,dma] Support non-wrapping, non-incrementing mode in FIFOs
Add support for the 'non-wrapping, non-incrementing' addressing
configuration for completeness. This means that each chunk is
transferred to a separate address, with all of the words within
a specific chunk being transferred to/from the same address.

Although this mode of operation is unlikely to be needed, the
DV previously prevented it using constraints because of
limitations in the FIFO model.

Change-Id: Ib99bcbedc5515ec2636f7034175ff5dea1d139c9
Signed-off-by: Adrian Lees <a.lees@lowrisc.org>