[rules,quality] Move module ID check top cc.bzl
The module ID check was previously running as an aspect. However this
can cause some problems because the aspect depends on the localtool
toolchain which sometimes creates weird issues leading to inscrutable
bazel error messages. With this commit, the module ID check is moved
to the cc.bzl so that it runs on all binaries created by
opentitan_{binary,test}. This also avoids the toolchain issue since
those are rules that can safely depend on the local toolchain.

On the technical side, this check is added the validation group. It
is enabled by default but can be disabled on the CLI. See
https://bazel.build/extending/rules#validation_actions for more info.

Change-Id: Ifbf71da614540eb610327dee01154a55a393691b
Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>