[dv] Downgrade more errors when building vcs simv with newer compilers
6d9e632e previously downgraded 'implicit-function-declaration' error into
warnings when building the vcs simv executable, which is an issue on newer
compiler releases which have changed the default severity for several checks.

int-conversion errors have also been observed for the same reason. An example is
given below.

Downgrade this error as well. (again, only while building the simv)

// rmapats.c:466:36: error: passing argument 1 of
// 'setChildClockWriteFuncAndPcode' makes integer from pointer without a cast
//  [-Wint-conversion]

Co-authored-by: Andreas Kurth <adk@lowrisc.org>
Change-Id: I30a7d5f6f335f51c175f6284e3b6870024445fb4
Signed-off-by: Harry Callahan <hcallahan@lowrisc.org>