[rules,rom] Mark legacy_rom_targets targets with compatible info
Currently, opentitan_binary creates a single target for all platforms
(cw310, cw340, sim_dv, etc). This is in contrast to opentitan_test which
creates one per platform. This creates a difficulty when building the
(test) ROM: we have a single target but we want to extract, say, the
CW310 VMEM for splicing. This is done by the legacy_rom_targets macro
which mimics per-platform targets which used to exists with the old
build rules.

Unfortunately, with multi-top this causes some issues. Taking the
test_rom target as an example: this target is marked as compatible
with all tops, but note that the Earlgrey target has a CW310 VMEM
whereas the Darjeeling does not. As a result, trying to build a
target created by legacy_rom_targets for the CW310 will try to extract
the CW310 VMEM, even for a Darjeeling build. Obviously, such a VMEM
does not exist so trying to expand this target somewhere will fail.

The solution adopted by this commit is to mark target created by
legacy_rom_targets with compatibility information based on the top.
This can be done by passing extend information to the macro so that
it knows, for example, that building a CW310 VMEM is only compatible
with Earlgrey. Fortunately, this problem had already arised with
opentitan_test so most of the infrastructure to figure this out is
already there.

Change-Id: I7d0fdf175756da6eb27d7f308f5a4180eaff94fb
Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>