[RIVOS][util,pyregtool] Fix compact multireg handling
This fixes handling of compact multiregs by considering each register
instance in such multiregs to be separate from one another, and
generating one register item per instance in the final pyreg/pydump
file, instead of grouping them under a single arrayed register item.

This is a more correct conversion because the register instance
definitions in a compact multireg may vary from one instance to the
next, so we can't use the definition of the first instance as an example
for all other instances like we do with regular multiregs.

An example of an IP that got converted incorrectly is the spidev, which
was missing all fields but one in its `CMD_FILTER` and `TPM_ACCESS`
registers. This fix ensures that there are as many registers generated
as necessary (respectively `CMD_FILTER_0` to `CMD_FILTER_7` and
`TPM_ACCESS_0` to `TPM_ACCESS_1`), and that all fields are present in
those (respectively `filter_0` to `filter_255` and `access_0` to
`access_4`).

Change-Id: I9e071de8b8ba855a69328c026206edb7528e200e
Signed-off-by: Alexandre Sollier <asollier@rivosinc.com>