[dv,dma] Improve coverage of ERROR_CODE register
Ensure that we've seen BusErr reported for each of the
ports, and as each of the source and destination of a
transfer.
Ensure that we've seen both source and address validation
checks fails on each of the ports being driven; the nature
of the validation differs amongst ports.
The errors signaled by the DMA shall be independent and
there is little/no point trying to achieve all permutations.

Change-Id: Ifabe37bab37934dd89637302a1261e3cf7889b40
Signed-off-by: Adrian Lees <a.lees@lowrisc.org>