[otp_ctrl] Add zeroization parameter to partitions
This commit adds a new zeroization parameter to the partition MMAP
which indicates whether its fuses can undergo the zeroization
procedure. We use a dedicated zeroization field as the zeroization
marker. The field is inserted after the digest when a partition is
labeled as zeroizable.

It further amends the partition generation script to only allow
partitions that are equipped with a digest to be zeroizable.

Autogen commands:
  - ./util/topgen.py -t hw/top_earlgrey/data/top_earlgrey.hjson
  - ./util/topgen.py -t hw/top_darjeeling/data/top_darjeeling.hjson

Reference:
  - https://github.com/chipsalliance/caliptra-ss/pull/518
  - https://github.com/chipsalliance/caliptra-ss/pull/579

Change-Id: I671c60af526eff45c04958a9b4b58ea2e0bc8ab2
Signed-off-by: Andrea Caforio <andrea.caforio@lowrisc.org>