[otp_ctrl/rtl] Zeroization detection
This commit inserts a zeroization check into buffered and unbuffered
partitions that is executed before the partition is
initialized. During this check the zeroization field is read out and
its value checked for the zeroization marker.

Detecting a zeroization incurs multiple side-effects with respect to
the periodic integrity/consistency checks.

Autogen commands:
  - ./util/topgen.py -t hw/top_earlgrey/data/top_earlgrey.hjson
  - ./util/topgen.py -t hw/top_darjeeling/data/top_darjeeling.hjson

Reference:
  - https://github.com/chipsalliance/caliptra-ss/pull/518
  - https://github.com/chipsalliance/caliptra-ss/pull/579

Change-Id: Ib53677b434bce3749e7e9146c2d83962ee999c19
Signed-off-by: Andrea Caforio <andrea.caforio@lowrisc.org>