[dv,dma] Exercise combinational responses
The DUT supports read/write traffic from/to combinational devices
(e.g. ROM/RAM connected directly to the TL-UL), which assert
`d_valid` in the same cycle that a transaction is accepted
('a_valid & a_ready').

Modify the DV environment to enable combinational responses.
This requires us to synchronize the notifications of A channel
and D channel activity within the scoreboard, which we now do
with minimal change to the current structure of the scoreboard;
collect both the A channel and the D channel items before
dispatching them into the existing checking logic.

Change-Id: Ibfb5c086494240becb96abad5b59a1d70285d0b0
Signed-off-by: Adrian Lees <a.lees@lowrisc.org>