[rtl,otp_ctrl] Add zeroization support for OTP
Add a new DAI command to cause the OTP to scrap partitions by writing
all ones to individual OTP items. Trigger the operation via the DAI
command and address CSRs.
Enhance the otp_mmap generation to mark individual partitions as
zeroizable, and error out on attempts to zeroize non-zeroizable
partitions.
Add a new 64-bit word at the end of any zeroizable partition, which
is used to mark the partition as zeroized, even if not all items are
actually zeroized.

This is to comply with FIPS requirements.

Change the darjeeling OTP mmap to make all partitions zeroizable.

Update the OTP Theory of Operation and Programmer's Guide docs.

Based on Caliptra's PRS below:

    chipsalliance/caliptra-ss#518
    chipsalliance/caliptra-ss#579

Change-Id: Ibcd54006b46201fa6a91e4d1e2c562571b7e58b3
Signed-off-by: Guillermo Maturana <matute@zerorisc.com>