[python] drop chipwhisperer dependency
This dependency is using the archive URL which are generated by GitHub
on the fly, which is not good for reliability and reproducibility
purpose. Releases should be used instead [0].

Upon on deeper inspection, however it seems that this dependency is not
used at all. This was used for SAM3X reset but this use is already
removed in [1]. So just drop it all-together.

Link: https://docs.github.com/en/repositories/working-with-files/using-files/downloading-source-code-archives [0]
Link: 46b156a66e1d0e4f7e651b55790347a8fae4fbc0 [1]
Change-Id: I6f940d2b6fd844143fac317ecebbffb85a118356
Signed-off-by: Gary Guo <gary.guo@lowrisc.org>