[dv,mbx] Lose the RAL for the SRAM port
Manage the TL agent for the SRAM connection directly because
it's a Device-mode TL agent and the concept of a RAL thus makes
no sense. The CIP layer assumes that there is a 1:1 mapping
between the tl_agent objects that it manages and the RAL names
with which it was supplied, but this is inappropriate when the
DUT is acting as a TL-UL host.

Change-Id: I1cd89767328abf03a3d702cca52dc0c3b420dbd9
Signed-off-by: Adrian Lees <a.lees@lowrisc.org>