[prim,fpv] Avoid prim_clock_mux2 using $isunknown in FPV
Jasper's behaviour is to make $isunknown(...) = 0, so this won't
change anything in the results of formal runs. But it gets rid of a
couple of warnings.

Since I was reasoning about this carefully, I've also tidied up the
comment a bit. In particular, the "including during reset" note is
bogus: there is no reset signal!

Change-Id: I672390205363de0ecfa451a58208e61f72f6dc23
Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>