[tlul,dv] Rephrase some sequences to make them simpler
These sequences were rather hard to understand (for me at least). They
were supposed to model the case where e.g. the host asserts a_valid on
the A channel and then changes the address without the original
transaction being accepted.

With the old sequences, we wanted to see valid but not ready and then
have a different value of e.g. a_addr on the next cycle where valid
was true.

I've simplified things a bit by saying that I'm only interested in
situations where the signal changes without valid dropping in the
middle.

My logic is that the aValidNotAccepted_S and dValidNotAccepted_S
sequences already capture the looser case where we're allowing valid
to drop in the middle, but this coverage target should probably want
the "most pesky" situation anyway, which is what you get from the
simpler and more understandable new SVA.

Change-Id: Id244467f41c26512040f034c5eb16f0700ccc14a
Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>