[pinmux,rtl] Tweak the FpvSecCmBusIntegrity_A assertion
This is analogous to the change we made in e.g. rom_ctrl (commit
0f7e1226574), where we change the assertion to actually be true.

In both cases, the original version of the assertion wasn't strictly
true because the alert receiver could stall the communication from the
alert sender, meaning that the alert_p signal would never go high.

Since we verify the alert system separately, this commit simplifies
things by just checking that we *ask* the alert system to do
something.

Change-Id: I50fd4b3ae9ad4c288472fcecc0e188864ebc6d78
Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>