[mbx] Refine imbx assertions
Write pointer advancing will not occur if the grant of the write
coincides with RoT-side use of the firmware reset (abort-clearing)
mechanism.

Remove assertion that was only checking the behavior of the DV;
in normal, correct use of the mailbox it is indeed not permissible
to write to WDATA whilst the mailbox is already busy, but CSR
tests do this and it would be preferable to leave them exercising
as many registers as possible.

In any case either the DUT should be robust against such writes or
the DV sequences should avoid them; assertions within the DUT are
for checking internal consistency and correct behavior of the
DUT logic itself.

Change-Id: Ie53cfe15cdadf49d700c9166ad9589d7f15f9dae
Signed-off-by: Adrian Lees <a.lees@lowrisc.org>