[test] Log pinmux attr config errors
Helps when debugging things such as opendrain not being available on
FPGAs. Without this log, it's not obvious from the error code which
attributes were not set.

Have to configure it for RISC-V only because this library is built for
the host by bindgen.

Change-Id: Icf140c437a531e535646d046f849581f56c018f7
Signed-off-by: James Wainwright <james.wainwright@lowrisc.org>
(cherry picked from commit cc75048554b3550c9eb02575a24e8c476dd4801b)