[opentitanlib] add tokio runtime
This is the beginning of the journey to remove the intrusion of MIO to
everywhere, and replacing them with `poll_*` alternatives and async.

To make one step at a time, we're not going to make everything sync, but
we will have a runtime available so we can have small regions of async code
and use `runtime().block_on()` to integrate them with existing sync code.

Change-Id: I9c86c2ba7839173e6d0f63a86523a4e256d1dd9b
Signed-off-by: Gary Guo <gary.guo@lowrisc.org>