[sw] Remove `compile_commands.json` generation script
This script has been broken for a while because of a problem with `bazel
aquery --config=riscv32` picking up Python files which are not supported
on the device side.

The databases that it generates are far too large (tens of GiBs) to be
usable given `clangd` loads them into memory.

Change-Id: I08943b4a9d1b1e74d7e83d379f7088ee6e11c98d
Signed-off-by: James Wainwright <james.wainwright@lowrisc.org>