[racl_ctrl,dv] Obey en_scb in racl_ctrl_scoreboard
The policy vector is asked to change by a register write, but this is
ignored when the scoreboard is disabled (by early returns in the
processs_tl_a_item and processs_tl_d_item tasks in
cip_base_scoreboard).

If we don't turn off watch_policies as well, csr_rw test fails because
we see the change to the output policy vector with watch_policies(),
but don't see the register write that caused it.

Change-Id: Ib7370bab94173cde29c2b019930fb68ab0de9904
Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>