[racl_ctrl,dv] Set priorities in write() calls
The existing code doesn't work properly if there is more than one
sequence item in flight on the agent, because the vseq has priority
-1.

Set the call in the base sequence to the default priority (100), since
it's doing something important. Set the call in the smoke sequence to
a lower priority (50), because it's intended to be a "background
task".

Change-Id: Ia46c9229b3c25d3ae020db0fcf00dcf996958208
Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>