[dv,mbx] Ensure STATUS coverage is collected
The STATUS register on the RoT interface is simply an alias
of SoC-side control and status bits so that DV has had no need
to read from that register; do so to ensure that the STATUS
register is observed and coverage is collected.

Check what we presently can, pending the later introduction of
loosely-timed interrupt/register predictions.

Change-Id: I42f59c73c1c5ac5f5ea8637b1e3db152ce526167
Signed-off-by: Adrian Lees <a.lees@lowrisc.org>