[csrng/dv] Fix paths to FSM-related asserts
dv/sva/csrng_assert_if.sv contains a list of FSM-related SVAs that need
to be turned off before running tests in which FSMs are forced to
invalid encodings/states. These paths were relative to the DUT, not to
the test environment, which resulted in these asserts not being found
when requested to be disabled (and re-enabled after the test).

Change-Id: Id9cf0ea0e11d56d36b2a309305d268d82cd7ac74
Signed-off-by: Florian Glaser <glaserf@lowrisc.org>