[topgen] Align memory naming with register blocks
Previously, memories would appear in templates (C, Rust, SV) as
<top>_<label>. However the label does not really have a meaning
before there isn't even a check for uniqueness and everything is
resolved internally using the instance name and interface name.

This PR changes the templates to use the <instance>_<int> name
instead of the label in templates. In addition, some template code
which was not using the top's helper class was rewritten.

Change-Id: Idaa6b09ff38d22c4ea675cb69b0e2341cd813f24
Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>