[racl_ctrl,dv] Teach the scoreboard to predict interrupts
The challenging thing here was teaching the prediction of when the
interrupt line should be high. This fundamentally depends on the
ERROR_LOG, INTR_ENABLE and INTR_TEST fields and I had to carefully
wire in tracking of write *requests* (and then clock to a stored
value).

Normally, this gets predicted nicely in the RAL, but the INTR_STATE
register is a bit fiddly, because it might get updated at the same
time as a bus access is trying to read it. The solution I've used here
is to check read responses, but allow the "new" value of the
INTR_STATE register as well.

Change-Id: I899860c2b69d96a58bd84b94799a3a61477f228b
Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>