[dif,clkmgr] Use the DT instead of hardcoded top information
Previously the DIF code was using a bunch of tricky macros to
extract the top specific information from register headers.
This commit rewrites this code using the DT information instead.
Note that this commit keeps the top specific enumeration in
the DIF header because some code relies on it but the DIF now
also provides a function to find the index for the DT.

Change-Id: Ibc4c1995e1a79c48c2f4868b93971c5c012c7ba8
Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>