[test_utils,clkmgr] Port to multitop
Two part remains top-specific: the code displays the clock names
for debugging but we don't have this information in the DT. There
is a function which does that in a semi-multitop fashion but in
the worst case will just display <unknown>.

Another top-independent part is the handling of external clocks.
Here it is not clear that the logic can be generalised so in
doubt it is still enclose in top-specific ifdefs.

Change-Id: I30e0e185b13e3091f76a020dd06aed7b3c6f2bda
Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>