[entropy_src/dv] Don't disable bucket health test SVAs in error test
These paths cannot be resolved due to the loop which causes tools to
throw errors and warnings. It turns out that we don't actually need
to disable these anyway: they only matter if we're glitching the esrng
FIFO at the very beginning of the ENTROPY_SRC hardware pipeline. This
FIFO will always contain valid data (there is no backpressure on this
FIFO and it starts filling up whenever the module is enabled).

This resolves lowRISC/OpenTitan#27810.

Change-Id: Ic0d2d331e65369a008217d51f553f90060301a49
Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>