[entropy_src/dv] Reduce max RNG rate for Darjeeling variant
The RNG in Darjeeling can run at a very high rate but the maximum rate
is slighly slower than what the RNG agent was producing for the max rate
sequence before. This commit, thus reduces the maximum RNG rate for the
Darjeeling configuration slightly to actually simulate real operating
conditions.

Change-Id: Ifae7b5ca7492c7bcdce5e742a93c3bd2e9813ae1
Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>