[otp_ctrl/rtl] Read zeroization marker always without checking integrity
Prior to this commit, a DAI read of the zeroization marker field of a
partition could result in an integrity error if the partition had been
zeroized but `otp_ctrl` hadn't been reset since the zeroization (which
would cause a reinitialization and, as part of that, the zeroization
state of the partition would have been updated in `otp_ctrl`).

This was problematic because SW couldn't just unconditionally go through
the zeroization markers of all partitions to determine which have been
zeroized and which haven't: if one partition had been zeroized without a
reset of `otp_ctrl`, reading the zeroization marker of that partition
would have caused a fatal integrity error, locking `otp_ctrl` up.

This commit fixes that by always using *raw* reads (i.e., reads that
ignore integrity checks) for the zeroization marker fields.

Change-Id: I92723c16613f4a0aa16c00f73d67e68e5f7bf34d
Signed-off-by: Andreas Kurth <adk@lowrisc.org>
Co-authored-by: Andrea Caforio <andrea.caforio@lowrisc.org>