[entropy_src/dv] Generalize entropy_src_intr_vseq for other symbol sizes
Previously, this virtual sequence was only working correctly for the
Earlgrey configuration using a 4-bit symbol size. This commit fixes
and generalizes the threshold computation and RNG generation for the
Darjeeling configuration. The functions stimulating the actual health
test failures are further improved to inject failures into individual
lanes rather always failing at the symbol level.

This is related to lowRISC/OpenTitan#27471.

Change-Id: I9c4d505f9ccfc20095cfcf156e934946414d7702
Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>