[hw,ip_templates] Add an OTP controller DT extension
In order to automatically generate the top specific SW readable OTP
partitions, we extend the OTP controller descriptor with an array of
partition information. Each entry in the array describes an OTP
partition that is still readable after the partition is locked.

The extension also defines a simple API to retrieve a partition
information from an OTP controller instance and a partition identifier.

Change-Id: I1bdda50556c20e0cbc0025a25dbadbdb4044887c
Signed-off-by: Samuel Ortiz <sameo@rivosinc.com>