[chip, dv] Made rstmgr_rst_cnsty_escalation working on Xcelium
This test pokes a consistency fault in rstmgr. Just before doing so,
the test writes to NMI_ENABLE register and expects NMI_STATE.Alert
to be raised. The way it works is by injecting fault into the path
that was registered inside the queue of proxies for rstmgr. In
return, u_alert_handler in top_earlgrey sends an escalation request
to u_rv_core_ibex.Inside the list of proxies, there are two paths
registered for fault injection for rstmgr as follows:

1) *u_rstmgr_aon.u_d0_spi_host_0.leaf_rst_o (generates a consistency
fault)
2) *u_rstmgr_aon.u_d0_spi_host_0.state_o (generates an FSM fault)

To grab the path from the list, the vseq chip_sw_fault_base_vseq
uses find_sec_cm_if_proxy function present in sec_cm_pkg. This
function contains a call to find_first() to find the first entry
that matches with the pattern *u_rstmgr_aon.u_d0_spi_host_0* inside
the list of proxies. Both VCS and Xcelium registers the above two
entries with different order. So, Xcelium gets the entry that turns
out to generate a non-consistency fault. But this test somehow
expects to generate a consistency fault. This is the reason why
u_alert_handler in top_earlgrey never sends an escalation request to
the core and as a result never writes to NMI_STATE.Alert.

Change-Id: Ib029d2f0af1c6f02ce6e918fefbecb1d0c8e34c7
Signed-off-by: Kinza Qamar <kqzaman@lowrisc.org>