[dv] Document cip_base_vseq
This commit re-orders cip_base_vseq.sv to use the out-of-block
definition style (except for the extra files that get included as
cip_base_vseq__*.svh).

The main reason for doing this was actually to carefully read and
document each of the tasks and functions in the class.

I've also dropped unneccessary "virtual" keywords and added a
smattering of "protected" and "local" keywords in the hope of avoiding
spaghetti.

Change-Id: I697f810d9933f1b7ee06449f1a9b8459e7adbd04
Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>