[opentitanlib] Add UART bitbang decoding sampling functionality
Implements additional sampling wrapper logic for the UART bitbanging
decoder, which takes uniform discrete samples despite the OpenTitanLib
`GpioMonitoring` interface providing the waveform as a list of edge
events and a start/end timestamp. This is self-contained so that anybody
wishing to use the UART bitbanging decoding interface can interface with
this wrapper to avoid duplicating the relevant parsing logic.

This performs the necessary optimisations to avoid needlessly sampling
whilst the UART is idle, but currently assumes that monitoring started
whilst the UART is idle.

Change-Id: Ic533f1321907a1f80513cee3aae2539ede568136
Signed-off-by: Alex Jones <alex.jones@lowrisc.org>