[sw,cryptolib] Follow OTBN style guide rule 5 and 9 related to bn.add/sub
This commit changes all the instructions in the p384 library that
violate rule 5 or 9 related to bn.add and bn.sub instructions.

After executing a bn.add or bn.sub instructions some sensitive state can
remain in the flags. For this reason we need to be careful with these
instructions and clear the flags in case they can contain sensitive
data.

Change-Id: I240cb393a53ac08debd8fae1ea2ecc606ba7add8
Signed-off-by: Hakim Filali <hfilali@lowrisc.org>