[crypto] Improve FI hardening of switch case statements
This PR improves the FI hardening of switch case statements. In each
case, the case condition is ORed into a common variable. After the
switch case statement, check if this common variable is equal to the
expected condition value. By using ORs, we can make sure that only
a single case statement was executed.

More details, including the assembly instructions of the improved
switch case statements,  can be found in lowRISC/opentitan#28203.

Change-Id: I00c955c1837f24996b818f19e5df4514cca265d8
Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>
Co-authored-by: Siemen Dhooghe <sdhooghe@google.com>