[rv_timer,dv] Fix subtly broken fork/join tree
With the previous version, we had something of this form:

  fork
    for (...) begin
      fork
        something_or_other();
      join_none
    end
    wait fork;
  join

This means the "wait fork" ends up starting in parallel with the thing
that was supposed to start sub-processes. This was all rather
confusing when I ran things with Xcelium because we end up with the
"wait fork" completing immediately, but then the something_or_other
process starting afterwards (no longer being waited for). This gets
followed by a "disable fork", leaving a process being killed when it
is trying to get a sequencer, causing a failure a cycle later.

I think the author probably intended something like this:

  begin
    for (...) begin
      fork
        something_or_other();
      join_none
    end
    wait fork;
  end

(where the outer begin/end is necessary to make this a single item in
a surrounding fork).

The only other change is to pull out a complicated expression, which
lets us indent it properly.

Change-Id: Idb9eba970ffafe54c3534b042d544df5ae695bb5
Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>