[rv_timer,dv] Sort out spinwait timeout in rv_timer_random_vseq
There are two changes. Firstly, the spinwait had a bizarre extra wait
after reading the right value! Reorder things to be a bit more
sensible.

Secondly, I realised that if delay is large and spinwait_timeout_ns is
small, we need to add to the timeout so that we actually wait long
enough to read the right value.

Change-Id: I3e4f56504a41ad039223e6aabea35aec6256329b
Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>