[entropy_src/rtl] Move alert counter clearing to end of HT window
This commit fixes an RTL bug that can occur when interfacing noise
sources operating at high rates. In particular, this commit ensures that
the point when alert counters get cleared is moved to right after
testing the last symbol belonging to the current health test window.

Prior to this commit, the clearing of the alert counters was controlled
by the main state machine which is operating on delayed versions of the
health test done / failed pulses where the amount of delay depends on
the exact configuration of the ENTROPY_SRC. As a result, it could happen
that the clearing pulse collided with the any_fail_pulse triggered
by the failure of one of the continuous health tests failing (repcnt and
repcnts). This could lead to continuous health test failures getting
ignored / not counting towards the total alert threshold.

This resolves lowRISC/OpenTitan#24159.

Change-Id: I913350ff028eb9e78f06afc0bc8a5e3a64897607
Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>