[pentest/CI] Set manual tag for fpga params
The firmware BUILD targets are opentitan_tests but these are not actual tests to run but instead are used to build the binaries to interface with the pentest framework.
Set the tags for both the fpga and silicon params to manual to not pull them into the nightly CI.

In addition, provide comments to warn users and remove the obsolete opentitan_binary targets.

Change-Id: I1323c07d1a22b6d094862c6147c4221115c1c621
Signed-off-by: Siemen Dhooghe <sdhooghe@google.com>