[sw,cryptolib] Only do instr count checks for OTBN status code OK
This commit moves down the instruction count checks for the ECC
ECDH finalize functions to after the OTBN status response check.
This is necessary since when a user input error occurs, OTBN doesnt
escalate to a fatal fault but returns a status code other than OK.
In this case the instruction count is not as expected and should not
be checked.

Change-Id: Iee7a784bff1d38c89a7f5edfacc2d290f37463ac
Signed-off-by: Hakim Filali <hfilali@lowrisc.org>