[bazel] Enable `--discard-locals` explicitly for linker
This flag is implicitly added when the toolchain is in "RISCVToolChain"
mode (distinct from Baremetal mode) [^1] which is enabled if `crt0.o` is
included in the toolchain or not [^2].

This may change in future builds of the toolchain, so we want to make
this flag explicit now.

The hash of the ROM and ROM_EXT ELFs are unchanged with this commit, so
I think there are no functional changes.

[^1]: https://github.com/llvm/llvm-project/blob/185b81e034ba60081023b6e59504dfffb560f3e3/clang/lib/Driver/ToolChains/RISCVToolchain.cpp#L166
[^2]: https://github.com/llvm/llvm-project/blob/e9f3be63d3928b24f3667d8aaadfbee9d325015f/clang/lib/Driver/ToolChains/BareMetal.cpp#L151

Change-Id: I778d2282aafe55bcdf62bbc80d76b533ac1bd920
Signed-off-by: James Wainwright <james.wainwright@lowrisc.org>