[topgen,rtl] Fix param handling for AlertSkewCycles for alert_handler
Previously AlertSkewCycles is only set for modules that have alert_connections.
However, the alert handler itself does not have alert_connections, so it was not
getting the AlertSkewCycles parameter set. Thus the alert handler was using the
default value of 1, causing an inconsistency if the parameter was set to a different
value.

This commit fixes this by adding AlertSkewCycles to the alert handler parameters.

Change-Id: Iaac183910e86b3530266d0c12114cf884a6fa33b
Signed-off-by: Robert Schilling <rschilling@rivosinc.com>